Finished implementation for Win32 and fixed major bug.
   The windows terminal I had was compiling to 32 bit!
   I had to make some fixes for 64 bit, all tests passing.

+++ src/coyote_linux_apple_common.h
@@ -143,7 +143,7 @@ ERR_RETURN:
 }
 
 static inline CoyMemoryBlock 
-coy_memory_allocate(int64_t minimum_num_bytes)
+coy_memory_allocate(intptr_t minimum_num_bytes)
 {
     CoyMemoryBlock result = {0};
     // TODO implement
